n7-sentinels/n7_sentinels/agent_runtime/service.py

import asyncio
import logging
import aiohttp
from .config import settings

logger = logging.getLogger("n7-sentinel.agent-runtime")

class AgentRuntimeService:
    """
    Agent Runtime Service.
    Responsibility: Handle registration, heartbeat, and configuration sync with Core.
    Ref: TDD Section 5.1 Sentinel Process Model
    """
    def __init__(self):
        self._running = False
        self._session = None
        self._token = None

    async def start(self):
        self._running = True
        logger.info("AgentRuntimeService started.")
        self._session = aiohttp.ClientSession()
        
        # Authenticate with Core
        await self._authenticate()
        
        # Start Heartbeat Loop
        asyncio.create_task(self._heartbeat_loop())

    async def stop(self):
        self._running = False
        if self._session:
            await self._session.close()
        logger.info("AgentRuntimeService stopped.")

    async def _authenticate(self):
        """
        Authenticates with Core to get JWT.
        """
        try:
            # TODO: Implement actual auth flow (e.g., using API Key or mTLS to exchange for JWT)
            # For now, simplistic placeholder
            logger.info("Authenticating with Core...")
            self._token = "mock-jwt-token"
        except Exception as e:
            logger.error(f"Authentication failed: {e}")

    async def _heartbeat_loop(self):
        """
        Sends periodic heartbeats to Core.
        """
        while self._running:
            try:
                if self._token:
                    headers = {"Authorization": f"Bearer {self._token}"}
                    # payload = {"status": "active", ...}
                    # async with self._session.post(f"{settings.CORE_API_URL}/agents/heartbeat", headers=headers) as resp:
                    #     if resp.status != 200:
                    #         logger.warning(f"Heartbeat failed: {resp.status}")
                    pass
            except Exception as e:
                logger.error(f"Heartbeat error: {e}")
            
            await asyncio.sleep(30)
